==> 00-setup.R <==
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
#                                                                             #
#  FEATURES AND FINDINGS OF THE METAPSY META-ANALYTIC RESEARCH DOMAIN         #
#  FOR PSYCHOLOGICAL TREATMENTS                                               #
#                                                                             #
#  0. Setup                                                                   #                                                      # nolint
#                                                                             #
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #


# STEP 1: Starting R from the Console -----------------------------------------
# Make sure R is running by typing 'R' into the Console below.
# Then hit Enter.

# STEP 2: Restore virtual environment -----------------------------------------
# Please note that restoring the environment can take 10-15 minutes!
renv::restore()

# STEP 3: Install REditorSupport ----------------------------------------------
# 1. Navigate to the "Extensions" pane (left sidebar)
# 2. Search for "REditorSupport"
# 3. Install "REditorSupport"

# For questions or problems with the environment, please contact
# Mathias Harrer (mathias.harrer@tum.de).
